refactor bins route to use new structure, moved applyEnforceHttps to config folder
