+++ models/user.js
@@ -40,4 +40,4 @@ UsersSchema.methods.toAuthJSON = function() {
   };
 };
 
-mongoose.model('Users', UsersSchema);
+mongoose.model('User', UsersSchema);
