+++ routes/api/index.js
@@ -2,6 +2,6 @@ const express = require('express');
 const router = express.Router();
 
 router.use('/users', require('./users'));
-// router.use('/bins', require('./bins'));
+router.use('/bins', require('./bins'));
 
 module.exports = router;
