+++ routes/api/users.js
@@ -2,7 +2,7 @@ const mongoose = require('mongoose');
 const passport = require('passport');
 const router = require('express').Router();
 const auth = require('../auth');
-const Users = mongoose.model('Users');
+const Users = mongoose.model('User');
 
 //POST new user route (optional, everyone has access)
 router.post('/', auth.optional, (req, res, next) => {
